address comments, leave test_helpers as static until further instruction
