+++ inc/az_ulib_ustream_base.h
@@ -329,6 +329,7 @@
 #include "az_ulib_config.h"
 #include "az_ulib_pal_os.h"
 #include "az_ulib_result.h"
+#include "az_ulib_base.h"
 
 #ifdef __cplusplus
 #include <cstddef>
@@ -339,21 +340,6 @@
 #endif /* __cplusplus */
 
 #include "azure/core/_az_cfg_prefix.h"
-
-#ifndef AZ_ULIB_USTREAM_FORWARD_BASE_H
-/**
- * @brief   Define offset_t with the same size as size_t.
- */
-typedef size_t offset_t;
-
-/**
- * @brief   Signature of the function to release memory passed to the ustream
- *
- * @param[in]   release_pointer       void pointer to memory that needs to be free'd
- *
- */
-typedef void (*az_ulib_release_callback)(void* release_pointer);
-#endif /* AZ_ULIB_USTREAM_FORWARD_BASE_H */
 
 /**
  * @brief   Forward declaration of az_ulib_ustream. See #az_ulib_ustream_tag for struct members.
