+++ tests/src/az_ulib_ustream_forward_mock_buffer.c
@@ -9,7 +9,6 @@
 
 static az_result _concrete_flush_result = AZ_OK;
 static az_result _concrete_read_result = AZ_OK;
-static az_result _concrete_get_size_result = AZ_OK;
 static az_result _concrete_dispose_result = AZ_OK;
 
 static offset_t current_position = 0;
@@ -62,15 +61,11 @@ static az_result concrete_read(
   return result;
 }
 
-static az_result concrete_get_size(az_ulib_ustream_forward* ustream_forward, size_t* const size)
+static size_t concrete_get_size(az_ulib_ustream_forward* ustream_forward)
 {
   (void)ustream_forward;
 
-  *size = 10;
-
-  az_result result = _concrete_get_size_result;
-  _concrete_get_size_result = AZ_OK;
-  return result;
+  return 10;
 }
 
 static az_result concrete_dispose(az_ulib_ustream_forward* ustream_forward)
@@ -98,7 +93,6 @@ az_ulib_ustream_forward* ustream_forward_mock_create(void)
 {
   _concrete_flush_result = AZ_OK;
   _concrete_read_result = AZ_OK;
-  _concrete_get_size_result = AZ_OK;
   _concrete_dispose_result = AZ_OK;
 
   return &USTREAM_FORWARD_COMPLIANCE_MOCK_BUFFER;
@@ -108,10 +102,5 @@ az_ulib_ustream_forward* ustream_forward_mock_create(void)
 void set_flush_result(az_result result) { _concrete_flush_result = result; }
 
 void set_read_result(az_result result) { _concrete_read_result = result; }
-
-void set_get_size_result(az_result result)
-{
-  _concrete_get_size_result = result;
-}
 
 void set_dispose_result(az_result result) { _concrete_dispose_result = result; }
